+++ tasks/serve.js
@@ -1,6 +1,5 @@
 var Promise = require('es6-promise').Promise;
 var log = require('./utils/log');
-var fs = require('./utils/fs');
 var helper = require('./utils/config-helper');
 var Browserify = require('./wrappers/browserify.js');
 var browserSync = require('browser-sync');
@@ -14,15 +13,53 @@ function initConfig(){
     pkg = config.pkg;
 }
 
+function getWatchOptions(){
+    var files = [];
+    var htmlPaths = [ globs.source.html.replace('*.{','**/*.{') ];
+    var stylesPaths = [globs.source.styles.replace('{.,*}','**').replace('!(_)','') ];
+    var scriptsPaths = [globs.source.scripts.replace('{.,*}','**') ];
+    if (paths.demo){
+        htmlPaths.push(globs.demo.html.replace('*.{','**/*.{'));
+        stylesPaths.push(globs.demo.styles.replace('{.,*}','**').replace('!(_)',''));
+        scriptsPaths.push(globs.demo.scripts.replace('{.,*}','**'));
+    }
+    var chokidarOptions = {
+        persistent: true
+    };
+    var callBack = function(task){
+        return function(event, file){
+            switch (event) {
+                case 'change':
+                    log.info(['Watch: File `' + file + '` has been changed, running build.' + task + '()'].join('\n'));
+                    build[task]().then(browserSync.reload);
+                    break;
+                default :
+                    log.info(' * ' + event + ' ' + file);
+                    break;
+            }
+        };
+    };
+    files.push({ match: htmlPaths, fn: callBack('html'), options: chokidarOptions });
+    files.push({ match: stylesPaths, fn: callBack('styles'), options: chokidarOptions });
+    if (helper.matches(config.build, ['browserify'])) {
+        browserifyWatch();
+    } else {
+        files.push({ match: scriptsPaths, fn: callBack('scripts'), options: chokidarOptions });
+    }
+    return files;
+}
+
 function start(options){
     initConfig();
     options = options || (config[config.serve]) || {};
+    //todo: test: does this now work as nodeApp??
     options.server = options.server || { baseDir : paths.target };
     return nodeApp(options).then(function(){
         log.warn('Server Started :');
-        if (!options.server && !options.proxy){
+        if (!options.server && !options.proxy){ //todo: now a problem?
             log.warn('caddy.config.js may be incorrect. please check');
         }
+        options.files = getWatchOptions();
         browserSync(options);
     });
 }
@@ -39,33 +76,10 @@ function nodeApp(options){
     });
 }
 
-function buildAndReload(task){
-    return function(){ build[task]().then(browserSync.reload); };
-}
-
-function watch(){
-    if (!config.build) return;
-    initConfig();
-    var paths = config.paths;
-    var fs = require('./utils/fs');
-    var htmlPaths = [ globs.source.html.replace('*.{','**/*.{') ];
-    var stylesPaths = [globs.source.styles.replace('{.,*}','**').replace('!(_)','') ];
+function browserifyWatch(){
+    new Browserify(globs.source.scripts, paths.target).watch(browserSync);
     if (paths.demo){
-        htmlPaths.push(globs.demo.html.replace('*.{','**/*.{'));
-        stylesPaths.push(globs.demo.styles.replace('{.,*}','**').replace('!(_)',''));
-    }
-    fs.watch(htmlPaths, [buildAndReload('html')]);
-    fs.watch(stylesPaths, [buildAndReload('styles')]);
-
-    if (helper.matches(config.build, ['browserify'])){
-        new Browserify(globs.source.scripts, paths.target).watch(browserSync);
-        if (paths.demo){
-            new Browserify(globs.demo.scripts, paths.target).watch(browserSync);
-        }
-    } else if (helper.matches(config.build, ['requirejs'])){
-        var scriptsPaths = [globs.source.scripts.replace('{.,*}','**') ];
-        paths.demo && scriptsPaths.push(globs.demo.scripts.replace('{.,*}','**'));
-        fs.watch(scriptsPaths,   [buildAndReload('scripts')]);
+        new Browserify(globs.demo.scripts, paths.target).watch(browserSync);
     }
 }
 
@@ -81,9 +95,7 @@ function adhoc(baseDir){
 }
 
 function run(args){
-    return start(args).then(function(){
-        watch();
-    }).catch(log.onError);
+    return start(args).catch(log.onError);
 }
 
 module.exports = {