+++ CHANGELOG.md
@@ -12,17 +12,20 @@
   * Removed `dist` path config option
   * Removed `site.` from Mustache/Jade variables i.e. `site.now` becomes `now`.
   * Added `copy` task to separate the task of copying images and server-config from building scripts etc.
-  * Added `package.json` to Mustache/Jade variables by default.
   * Added `html-min` as a separate build task.
   * Added `bower` as a separate `release` option.
   
 todo:
-  * speed, speed, speed
   * don't minify in dev mode
+  * test 'new'
+  * test 'init'
+  * test 'release'
+  * test in windows
+  * speed, speed, speed
   * document how to test a specific .spec file (fit + fdescribe)
+  * document `*.requirejs.js` into a `use-case doc` to reduce confusing default code
   * make tasks only use wrappers . no fs utils etc
   * move all logging into tasks and out of wrapper/util files
-  * move `*.requirejs.js` into a `use-case doc` to reduce confusing default code
   * windows "when I save column.js=, it does refresh, but it doesn't always reflect the changes"
   * standardise input to all tasks i.e. `src, dest, options`
   * //kill test tdd properly